Add files via upload
A program written to perform statistical calculations on a range of resistors inputed by the user. The program is built to allow for user to input the resistor values in a  robust manner and to allow user to pick which statistical operation will be carried out. It then gives the user the option to restart without closing it.